+++ icevision/models/mmdet/common/bbox/two_stage/model.py
@@ -4,6 +4,7 @@ from icevision.imports import *
 from mmcv import Config
 from mmdet.models import build_detector
 from mmcv.runner import load_checkpoint
+from icevision.models.mmdet.utils import *
 
 
 def model(
@@ -17,7 +18,13 @@ def model(
     if isinstance(cfg, (str, Path)):
         cfg = Config.fromfile(str(cfg))
 
-    cfg.model.roi_head.bbox_head.num_classes = num_classes - 1
+    # Sparse-RCNN has a list of bbox_head whereas Faster-RCNN has only one
+    if isinstance(cfg.model.roi_head.bbox_head, list):
+        for bbox_head in cfg.model.roi_head.bbox_head:
+            bbox_head["num_classes"] = num_classes - 1
+    else:
+        cfg.model.roi_head.bbox_head.num_classes = num_classes - 1
+
     if weights_path is not None:
         cfg.model.pretrained = None
 
@@ -25,5 +32,7 @@ def model(
 
     if weights_path is not None:
         load_checkpoint(_model, str(weights_path))
+
+    _model.param_groups = MethodType(param_groups, _model)
 
     return _model