+++ icevision/models/mmdet/utils.py
@@ -1,0 +1,27 @@
+__all__ = ["param_groups"]
+
+from icevision.imports import *
+from icevision.utils import *
+from mmdet.models.detectors import *
+
+
+def param_groups(model):
+    body = model.backbone
+
+    layers = []
+    layers += [nn.Sequential(body.conv1, body.bn1)]
+    layers += [getattr(body, l) for l in body.res_layers]
+    layers += [model.neck]
+
+    if isinstance(model, SingleStageDetector):
+        layers += [model.bbox_head]
+    elif isinstance(model, TwoStageDetector):
+        layers += [nn.Sequential(model.rpn_head, model.roi_head)]
+    else:
+        raise RuntimeError(
+            "{model} must inherit either from SingleStageDetector or TwoStageDetector class"
+        )
+
+    _param_groups = [list(layer.parameters()) for layer in layers]
+    check_all_model_params_in_groups2(model, _param_groups)
+    return _param_groups