+++ icevision/models/torchvision/mask_rcnn/prediction.py
@@ -3,16 +3,19 @@ __all__ = ["predict", "predict_dl", "convert_raw_prediction", "convert_raw_predi
 from icevision.imports import *
 from icevision.utils import *
 from icevision.core import *
+from icevision.data import *
 from icevision.models.utils import _predict_dl
+from icevision.models.torchvision.mask_rcnn.dataloaders import *
 from icevision.models.torchvision.faster_rcnn.prediction import (
     convert_raw_prediction as faster_convert_raw_prediction,
 )
 
 
 @torch.no_grad()
-def predict(
+def _predict_batch(
     model: nn.Module,
     batch: List[torch.Tensor],
+    records: Sequence[BaseRecord],
     detection_threshold: float = 0.5,
     mask_threshold: float = 0.5,
     device: Optional[torch.device] = None,
@@ -24,8 +27,25 @@ def predict(
     raw_preds = model(*batch)
     return convert_raw_predictions(
         raw_preds=raw_preds,
+        records=records,
         detection_threshold=detection_threshold,
         mask_threshold=mask_threshold,
+    )
+
+
+def predict(
+    model: nn.Module,
+    dataset: Dataset,
+    detection_threshold: float = 0.5,
+    device: Optional[torch.device] = None,
+) -> List[Prediction]:
+    batch, records = build_infer_batch(dataset)
+    return _predict_batch(
+        model=model,
+        batch=batch,
+        records=records,
+        detection_threshold=detection_threshold,
+        device=device,
     )
 
 
@@ -33,7 +53,7 @@ def predict_dl(
     model: nn.Module, infer_dl: DataLoader, show_pbar: bool = True, **predict_kwargs
 ):
     return _predict_dl(
-        predict_fn=predict,
+        predict_fn=_predict_batch,
         model=model,
         infer_dl=infer_dl,
         show_pbar=show_pbar,
@@ -42,30 +62,40 @@ def predict_dl(
 
 
 def convert_raw_predictions(
-    raw_preds: List[dict], detection_threshold: float, mask_threshold: float
+    raw_preds: List[dict],
+    records: Sequence[BaseRecord],
+    detection_threshold: float,
+    mask_threshold: float,
 ):
     return [
         convert_raw_prediction(
             raw_pred=raw_pred,
+            record=record,
             detection_threshold=detection_threshold,
             mask_threshold=mask_threshold,
         )
-        for raw_pred in raw_preds
+        for raw_pred, record in zip(raw_preds, records)
     ]
 
 
 def convert_raw_prediction(
-    raw_pred: dict, detection_threshold: float, mask_threshold: float
+    raw_pred: dict,
+    record: Sequence[BaseRecord],
+    detection_threshold: float,
+    mask_threshold: float,
 ):
-    preds = faster_convert_raw_prediction(
-        raw_pred=raw_pred, detection_threshold=detection_threshold
+    pred = faster_convert_raw_prediction(
+        raw_pred=raw_pred, record=record, detection_threshold=detection_threshold
     )
 
-    above_threshold = preds["above_threshold"]
+    above_threshold = pred.detection.above_threshold
     masks_probs = raw_pred["masks"][above_threshold]
     masks_probs = masks_probs.detach().cpu().numpy()
     # convert probabilities to 0 or 1 based on mask_threshold
     masks = masks_probs > mask_threshold
     masks = MaskArray(masks.squeeze(1))
 
-    return {"masks": masks, **preds}
+    pred.pred.add_component(MasksRecordComponent())
+    pred.detection.set_masks(masks)
+
+    return pred