+++ icevision/parsers/parser.py
@@ -2,9 +2,9 @@ __all__ = ["ParserInterface", "Parser"]
 
 from icevision.imports import *
 from icevision.utils import *
+from icevision.utils.code_template import *
 from icevision.core import *
 from icevision.data import *
-from icevision.parsers.mixins import *
 
 
 def camel_to_snake(name):
@@ -21,7 +21,7 @@ class ParserInterface(ABC):
         pass
 
 
-class Parser(ClassMapMixin, ImageidMixin, SizeMixin, ParserInterface, ABC):
+class Parser(ParserInterface, ABC):
     """Base class for all parsers, implements the main parsing logic.
 
     The actual fields to be parsed are defined by the mixins used when
@@ -41,49 +41,55 @@ class Parser(ClassMapMixin, ImageidMixin, SizeMixin, ParserInterface, ABC):
     """
 
     def __init__(
-        self, class_map: Optional[ClassMap] = None, idmap: Optional[IDMap] = None
+        self,
+        record,
+        class_map: Optional[ClassMap] = None,
+        idmap: Optional[IDMap] = None,
     ):
-        self.class_map = class_map or ClassMap()
-        if class_map is None:
-            self.class_map.unlock()
+        # self.class_map = class_map or ClassMap()
+        # if class_map is None:
+        #     self.class_map.unlock()
+        self._record = record
         self.idmap = idmap or IDMap()
 
     @abstractmethod
     def __iter__(self) -> Any:
         pass
 
+    @abstractmethod
+    def parse_fields(self, o, record: BaseRecord) -> None:
+        pass
+
+    def create_record(self) -> BaseRecord:
+        return deepcopy(self._record)
+
     def prepare(self, o):
         pass
 
-    def record_class(self) -> BaseRecord:
-        return create_mixed_record(self.record_mixins())
-
     def parse_dicted(self, show_pbar: bool = True) -> Dict[int, RecordType]:
-
-        Record = self.record_class()
         records = {}
 
         for sample in pbar(self, show_pbar):
             try:
                 self.prepare(sample)
-                true_imageid = self.imageid(sample)
-                imageid = self.idmap[true_imageid]
+                # TODO: Do we still need idmap?
+                true_record_id = self.record_id(sample)
+                record_id = self.idmap[true_record_id]
 
                 try:
-                    record = records[imageid]
+                    record = records[record_id]
                 except KeyError:
-                    record = Record()
+                    record = self.create_record()
+                    # HACK: fix record_id (needs to be transformed with idmap)
+                    record.set_record_id(record_id)
+                    records[record_id] = record
 
                 self.parse_fields(sample, record)
 
-                # HACK: fix imageid (needs to be transformed with idmap)
-                record.set_imageid(imageid)
-                records[imageid] = record
-
             except AbortParseRecord as e:
                 logger.warning(
-                    "Record with imageid: {} was skipped because: {}",
-                    true_imageid,
+                    "Record with record_id: {} was skipped because: {}",
+                    true_record_id,
                     str(e),
                 )
 
@@ -113,10 +119,6 @@ class Parser(ClassMapMixin, ImageidMixin, SizeMixin, ParserInterface, ABC):
         # Returns
             A list of records for each split defined by `data_splitter`.
         """
-
-        # Hack to define the class of the mixed_record in the local namespace. This is required for pickeling.
-        Record = self.record_class()
-
         if self._check_path(cache_filepath):
             logger.info(
                 f"Loading cached records from {cache_filepath}",
@@ -138,7 +140,7 @@ class Parser(ClassMapMixin, ImageidMixin, SizeMixin, ParserInterface, ABC):
 
                 all_splits_records.append(split_records)
 
-            self.class_map.lock()
+            # self.class_map.lock()
             if cache_filepath is not None:
                 pickle.dump(all_splits_records, open(Path(cache_filepath), "wb"))
 
@@ -150,6 +152,19 @@ class Parser(ClassMapMixin, ImageidMixin, SizeMixin, ParserInterface, ABC):
         return ["def __iter__(self) -> Any:"] + templates
 
     @classmethod
-    def generate_template(cls):
-        for template in cls._templates():
-            print(f"{template}")
+    def generate_template(cls, record):
+        record_builder_template = record.builder_template()
+
+        template = CodeTemplate()
+        template.add_line(f"class MyParser({cls.__name__}):", 0)
+        template.add_line(f"def __init__(self, record):", 1)
+        template.add_line(f"super().__init__(record=record)", 2)
+        template.add_line(f"def __iter__(self) -> Any:", 1)
+        template.add_line(f"def __len__(self) -> int:", 1)
+        # template.add_line("def create_record(self) -> BaseRecord:", 1)
+        # template.add_line(f"return {record.__class__.__name__}({components_names})", 2)
+        template.add_line("def record_id(self, o) -> Hashable:", 1)
+        template.add_line("def parse_fields(self, o, record):", 1)
+        template.add_lines(record_builder_template, 2)
+
+        template.display()