+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
+**BREAKING:** API Refactor
+
+## Changed
+**Breaking:** Renamed tasks `default,detect,classif` to `common,detection,classification`
+**Breaking:** Renamed `imageid` to `record_id`
+
+## [0.6.0b1]
 ### Added
 - mmdetection models